+++ .eslintrc.json
@@ -39,9 +39,7 @@
         "regescape": true,
         "unregescape": true,
         "model": true,
-        "kwicResults": true,
         "wordPictureResults": true,
-        "statsResults": true,
         "view": true,
         "CSV": true,
         "locationSearch": true,
