+++ app/scripts/services.js
@@ -273,10 +273,10 @@ korpApp.factory("searches", [
             }
 
             kwicSearch(cqp) {
+                // TODO maybe merge these to one event, including
+                // future events for word picture and other kinds of tabs
                 $rootScope.$emit("kwic_make_request", cqp)
-                if (window.statsResults) {
-                    statsResults.makeRequest(cqp)
-                }
+                $rootScope.$emit("stats_make_request", cqp)
             }
 
             lemgramSearch(word, type) {
